Build tool integration
