--- model/Direction.java
@@ -1,6 +1,0 @@
-package model;
-//a játékos/gép által végzett mozgás végrehajtásához segítség
-
-public enum Direction{
-	DOWN, LEFT, RIGHT
-}
