--- model/FieldState.java
@@ -1,6 +1,0 @@
-package model;
-//a tábla mezőinek állapotkövetésére
-
-public enum FieldState{
-	EMPTY, OCCUPIED
-}
