--- model/PieceColor.java
@@ -1,6 +1,0 @@
-package model;
-//blokk színei
-
-public enum PieceColor{
-	BLUE, GREEN, RED, YELLOW, ORANGE, CYAN, PURPLE, DEFAULT
-}
