--- model/PieceType.java
@@ -1,5 +1,0 @@
-package model;
-
-public enum PieceType{
-	STRAIGHT, SQUARE, TRIANGLE, RTURN, LTURN, LEFTL, RIGHTL
-}
