+++ src/model/Direction.java
@@ -1,0 +1,6 @@
+package model;
+//a játékos/gép által végzett mozgás végrehajtásához segítség
+
+public enum Direction{
+	DOWN, LEFT, RIGHT
+}
