+++ src/model/FieldState.java
@@ -1,0 +1,6 @@
+package model;
+//a tábla mezőinek állapotkövetésére
+
+public enum FieldState{
+	EMPTY, OCCUPIED
+}
