+++ src/model/PieceColor.java
@@ -1,0 +1,6 @@
+package model;
+//blokk színei
+
+public enum PieceColor{
+	BLUE, GREEN, RED, YELLOW, ORANGE, CYAN, PURPLE, DEFAULT
+}
