+++ src/model/PieceType.java
@@ -1,0 +1,5 @@
+package model;
+
+public enum PieceType{
+	STRAIGHT, SQUARE, TRIANGLE, RTURN, LTURN, LEFTL, RIGHTL
+}
